+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Added
 - color picker to patternlib
+- organization authority detail
 
 ### Fixed
 - authority year range filter